refactor unit tests into /unit
